Update CMakeLists.
fix the MAKE command needed for creating the kernel, ninja needs to
explicitly use make where $(MAKE) will be used.

Add AddRunTarget function to apps/CMakeLists.txt and remove
simulate add_custom_target as its not needed.

The reason AddRunTarget was needed is because with make we can use
have an "app" parameter:

  make simulate app=helloworld

But that doesn't work for ninja instead the function AddRunTarget is used
to create a specific target for each application.